add gps-watchdog
gps-watchdog is a small tool that will connect to gpsd as a client using
libgps.  If no data is received for a configurable timeout, a
configurable systemd service will be killed via 'systemctl kill'.

The idea is that this will restart gpsd, and some ExecStartPre of the
gpsd service is then re-initializing the gps receiver hardware.